Squashed commit of the following:
commit ba757212b08972b2cbfa560efac38e6efa707a2a
Author: sanjeev.bhandari <sanjeev.bhandari@tai.com.np>
Date:   Fri Jul 12 23:29:40 2024 +0545

Added functionality in web and updated README

commit 3ca394c5d4d26bc40c9fc019712df3c564e7eae1
Author: sanjeev.bhandari <sanjeev.bhandari@tai.com.np>
Date:   Fri Jul 12 23:19:23 2024 +0545

    Updating README.md

commit a41efae01f0f4f71f18e661f38a32dccfe96a827
Author: sanjeev.bhandari <sanjeev.bhandari@tai.com.np>
Date:   Fri Jul 12 23:10:34 2024 +0545

    added completed tasks

commit bdf9d705cdecceb59412cb6033764ced697e6bc1
Author: sanjeev.bhandari <sanjeev.bhandari@tai.com.np>
Date:   Fri Jul 12 23:09:20 2024 +0545

    added completed tasks

commit a93028b0d82a8f2ad100accc0a7705837ae3c615
Author: sanjeev.bhandari <sanjeev.bhandari@tai.com.np>
Date:   Fri Jul 12 22:20:18 2024 +0545

    Fixed data entry and logout view

+++ todo_task/myapp/views.py
@@ -1,23 +1,25 @@
 from typing import Any
 
 from django.contrib import messages
-from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.db.models.query import QuerySet
 from django.shortcuts import get_object_or_404, redirect, render
+from django.utils.decorators import method_decorator
 from django.views import generic
-
 from task_api.models import TodoTask
-from .forms import LoginForm, SignUpForm
 
 
+@method_decorator(login_required, name="dispatch")
 class GenericListView(generic.ListView):
     template_name = "myapp/index.html"
     context_object_name = "todo_list"
 
     def get_queryset(self) -> QuerySet[Any]:
-        return TodoTask.objects.order_by("updated")
+        return TodoTask.objects.filter(
+            completed=False, user=self.request.user
+        ).order_by("updated")
 
 
 def signup_view(request):
@@ -32,14 +34,19 @@ def signup_view(request):
         elif User.objects.filter(username=username).exists():
             messages.error(request, f"A user with username {username} already exists.")
         elif password1 != password2:
-            messages.error(request, "Password and Confirmation password must be the same.")
+            messages.error(
+                request, "Password and Confirmation password must be the same."
+            )
         else:
-            user = User.objects.create_user(username=username, email=email, password=password1)
+            user = User.objects.create_user(
+                username=username, email=email, password=password1
+            )
             messages.success(request, "Your account has been created successfully.")
             login(request, user)
             return redirect("todos:login")
 
     return render(request, "account/signup.html")
+
 
 def login_view(request):
     if request.method == "POST":
@@ -56,10 +63,12 @@ def login_view(request):
 
     return render(request, "account/login.html")
 
+
 @login_required
 def logout_view(request):
     logout(request)
     return redirect("todos:login")
+
 
 @login_required
 def add(request, *args, **kwargs):
@@ -68,13 +77,20 @@ def add(request, *args, **kwargs):
     if desc is None:
         desc = task
 
-    TodoTask.objects.create(task=task, desc=desc)
+    # Create a TodoTask associated with the authenticated user
+    TodoTask.objects.create(user=request.user, task=task, desc=desc)
+
     return redirect("todos:index")
+
 
 @login_required
 def delete(request, *args, **kwargs):
     id = kwargs.get("todo_id")
     todo = get_object_or_404(TodoTask, pk=id)
+
+    if todo.user != request.user:
+        messages.error(request, "You are not authorized to delete this task.")
+        return redirect("todos:index")
 
     todo.delete()
     return redirect("todos:index")
@@ -88,6 +104,22 @@ def update(request, *args, **kwargs):
         completed = True
     todo = get_object_or_404(TodoTask, pk=id)
 
+    if todo.user != request.user:
+        messages.error(request, "You are not authorized to update this task.")
+        return redirect("todos:index")
+
     todo.completed = completed
     todo.save()
     return redirect("todos:index")
+
+
+@method_decorator(login_required, name="dispatch")
+class CompletedTaskListView(generic.ListView):
+    template_name = "myapp/completed_task.html"
+    context_object_name = "todo_list"
+    paginate_by = 10  # Number of tasks per page
+
+    def get_queryset(self):
+        return TodoTask.objects.filter(completed=True, user=self.request.user).order_by(
+            "updated"
+        )
